IngredientCard element
